+++ go.mod
@@ -1,5 +1,3 @@
 module github.com/growthbook/growthbook-golang
 
 go 1.17
-
-require github.com/franela/goblin v0.0.0-20211003143422-0a4f594942bf
